requirements.txt
blinker==1.9.0
cached-property==2.0.1
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
flask==3.1.1
flask-sqlalchemy==3.1.1
idna==3.10
inflection==0.5.1
itsdangerous==2.2.0
jinja2==3.1.6
markupsafe==3.0.2
mypy-extensions==1.1.0
python-dotenv==1.1.1
requests==2.32.4
sqlalchemy==2.0.41
sqlalchemy-orm==1.2.10
typing-extensions==4.14.0
typing-inspect==0.9.0
urllib3==2.5.0
werkzeug==3.1.3
